uploading a file works on frontend, backend multipart receiving works

+++ server/createDB.sql
@@ -8,7 +8,7 @@ CREATE TABLE videos(
     id INT PRIMARY KEY NOT NULL AUTO_INCREMENT,
     name VARCHAR(255) NOT NULL,
     size INT,
-    duration INT,
+    duration FLOAT,
     date_recorded DATETIME,
     url varchar(255) NOT NULL
 );
